requirements.txt
asgiref==3.5.2
autopep8==2.0.0
beautifulsoup4==4.11.1
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.1.3
django-admin-volt==1.0.4
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-crispy-forms==1.14.0
django-filter==22.1
django-isbn-field==0.5.3
django-jet-reboot==1.3.3
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-reverse-admin==2.9.6
django-tables2==2.4.1
djangorestframework==3.14.0
mysqlclient==2.1.1
phonenumbers==8.13.2
Pillow==9.3.0
pycodestyle==2.10.0
python-stdnum==1.18
pytz==2022.6
qrcode==7.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
